move session-related code to new session module
