+++ dill/__init__.py
@@ -24,13 +24,17 @@ except: # pragma: no cover
 
 
 from ._dill import (
-    dump, dumps, load, loads, dump_module, load_module, load_module_asdict,
-    dump_session, load_session, Pickler, Unpickler, register, copy, pickle,
-    pickles, check, HIGHEST_PROTOCOL, DEFAULT_PROTOCOL, PicklingError,
-    UnpicklingError, HANDLE_FMODE, CONTENTS_FMODE, FILE_FMODE, PickleError,
-    PickleWarning, PicklingWarning, UnpicklingWarning,
+    Pickler, Unpickler,
+    check, copy, dump, dumps, load, loads, pickle, pickles, register,
+    DEFAULT_PROTOCOL, HIGHEST_PROTOCOL, CONTENTS_FMODE, FILE_FMODE, HANDLE_FMODE,
+    PickleError, PickleWarning, PicklingError, PicklingWarning, UnpicklingError,
+    UnpicklingWarning,
 )
-from . import source, temp, detect
+from .session import (
+    dump_module, load_module, load_module_asdict,
+    dump_session, load_session # backward compatibility
+)
+from . import detect, session, source, temp
 
 # get global settings
 from .settings import settings
